--- BullsAndCows/src/main/java/com/brainstars/bullsandcows/controllers/LoginController.java
@@ -1,13 +1,0 @@
-package com.brainstars.bullsandcows.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-  @GetMapping("/login")
-  public String showLogin() {
-    return "login";
-  }
-}
-
